refactor: merge checksum tests into one only file
